Add helper for tests and change all options-tests
After I added 4 Tests to options.js, I startet to write a little helper.
It provides a way to write tests faster. You don't have to deal with
array sizes any more and you can insert a new test at the beginning
without the need to change every other test after it.
If there are differences between the tests and 'the reality', there are
more details in the output to find the error (this also helps to write
new tests).
If ever needed, it should be more easy to change the test environment,
e.g. from expresso to nodeunit.

The helper tests for the very same as the assert-calls before - with
one exception: per default, the sequence of the tests isn't relevant
anymore, only "line" and "message" matters.