Extend option regexdash to detect dash in the middle of a set
The regexdash option prevents from throwing an error, if a dash is
the last character in a character set, but not, if it's the first
character.

Before this commit, sets like [a-c-e] (which is the same as
[a-c\-e]) are allowed, even if regexdash was not used.

As mentioned in #258 (and #159 and #83) a warning should be thrown
only, if the dash is in the middle of a set without being part of
a character range.

Closes GH-258
Closes GH-83