Parse directives more correct
This patch changes the way, directives are parsed - all strings at the
beginning of a scope are treated as string. Kind of asi is implemented
there (still not perfect but it should fit the most needs) - but even
with option asi on, there will be warnings (as proposed in #300).

Now it's allowed to add comments and other directives..
It fixes #168