Fixed incorrect column numbers when detecting whitespace.
This patch fixes GH-251 by passing the left token instead of
nexttoken to the warning function. This provides us with the
correct column count.

In addition to that, this patch expands our TestRunner to
support checking values other than line numbers.

Fixes GH-251.

(This is a squash commit of GH-324)