Correctly identify duplicate object properties.
JSHint was incorrectly identifying duplicate properties because of
numerous unsafe assumptions such as that setters always follow getters
and others.

This patch removes these assumptions and makes sure that we understand
that 'get x, set x is okay' but 'get x, x' is not.

+ Tests for duplicate members and a regression test for the
  aforementioned bug.

This is a response to #382 and a faulty merge
63382c7fcc17d5bfe86441f0ec5d8ef651138eff.