Move "x is not defined" check into the body of try/catch.
Fixes GH-394. We queue "x is not defined" warnings because sometimes
variables are defined after they are used (this is valid in JavaScript).
But if they were not defined at all, we apply the queued warning. This
was breaking JSHint when there were too many errors: our code was
outside of the try/catch block but still called quit() that throws
an exception.

This commit fixes that by moving the code into the body of try/catch.