Support compatibility with inverted JSLint options
Several of the options inherited from JSLint have retained their
original names but inverted their meanings.  Since the "/*jslint" option
syntax is still supported, running JSHint on a project using the
"/*jslint" option syntax with these options produces a result that is
opposite to the intentions of the developer.

To support this use case and increase compatibility with JSLint, invert
such option values when encountered with the "/*jslint" option syntax.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>