Make jshint-rhino.js runnable
Add a shebang to jshint-rhino.js to tell the shell how to execute the
file and mark it executable after build.

Note1:  Ignore the return code of chmod in case the user is on a
filesystem which does not support setting execute permission.
Note2:  Make use  of /usr/bin/env to catch rhino anywhere in the PATH.

Also reformat the rhino build target slightly to match the other targets
(replace chained '&&' with separate shell commands).  Combine multiple
invocations of cat into one.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>