Don't warn when Object is used as a function.
Like Boolean, Number, and String, Object can be called as a function,
which provides scripts with access to ECMAScript's internal ToObject
logic. If passed an object, return it. If passed a primitive, return
a new Number, Boolean, or String as appropriate (or new Object for
null/undefined). This is useful primarily when implementing shims for
new ECMAScript functionality that is specified to use the ToObject
internal method.

Closes GH-392.