Make our parser to be smarter about /= and regexs.
The ridiculously long regular expression that produces tokens was
confused about differences between '/=' and regular expressions that
start with '='. Basically it didn't know that the latter exist.

This patch fixes the problem using the simplest (but still dumb)
solution. It makes sure that the token for /= is not followed by
non-whitespace characters and '/[gim]'. I hope this will catch
most regular expressions out there.

Closes GH-536.