Fixed a bug where new max* options were not recognized.
So while publishing r11 I noticed that new max* options (maxparams,
etc.) were not recognized when set using JSHint comments. That was
because we forgot to update the code that checks for numeric values.
This patch fixes that and adds a regression test (just in case).

Will have to move r11 pointer to this commit instead of previous one.