Fix a bug where maxparams was failing on functions with no parameters.
verifyMaxParametersPerFunction expected the result of functionparams()
to be always an array. This is a bad assumption because when there are
no parameters, this function returns nothing.

Just changing functionparams() to return empty array was breaking
some ES5 parsing code so I decided to make
verifyMaxParametersPerFunction to be aware of possible 'undefined'.

Closes GH-642.