Don't assume 'strict' for 'node' environments.
We don't need to warn people about using "use strict" outside
of a function scope in Node environments but my previous naive
implementation was achieving this by simply turning 'globalstrict'
on. This doesn't really work because it 'globalstrict' implies
'strict' which  is not always wanted. This patch fixes it.

Fixes GH-721.