Added support of /*exported … */ directive
This patch introduces a new `/*exported … */` directive that is intended for use with the `unused` option. This directive will suppress warnings about unused function declarations and expressions if they are defined in the global scope but not used.

The following should run without warnings:

    /*exported cats */
    function cats() {}

If the function declaration or expression is not in the global scope, it should throw a warning:

    /*exported cats */
    function isCat() {
        var cats; // Warning: 'cats' is defined but never used.
    }
    isCat();

References:

Closes GH-659