Lexer refactoring.
This patch transforms lexer from module level functions that
operate with module-level variables to a normal JavaScript
object with its own state and so on. This means that we don't
need to explicitly reset variables between executions since
`new Lexer(...)` guarantees that the state will be new on every
run.

Also Lexer is an EventEmitter so we don't need to pass warningAt,
errorAt and quit functions. We just need to subscribe to the
warning, error and fatal Lexer events and call appropriate
functions from our handlers (see jshint.js for an example).