Better support for comma operator.
Before when JSHint was parsing expressions inside the prefix parens
it allowed for only one expression. This patch changes the function
to allow for multiple comma-separated expressions. This helps to deal
with tons of false-positive messages that we were spitting out before
(check out the reports for lodash and prototype.js for example).

Closes GH-56.