Improved prereg detection.
prereg is a flag we set whenever the next token can possibly start
a regular expression. It is used to disambiguate between the division
operator (/) and regular expressions. Before, we were whitelisting
chars that can be followed by a regular expression but this commit
changes the behavior to blacklisting "bad" tokens.

This resolved mysterious false positives when parsing Prototype.js.

Closes GH-759.
Closes GH-760.