Clear implied variables even when undef is false.
We have a pretty hacky code that clears implied variables that
were defined later on. The problem is that code is called only
when the undef option is on. So this means that when it's off
JSHINT.data().implieds array can be populated with false positive
data.

This patch changes the code so that it always collects and
clears undefined variables but checks the undef option when
it's ready to warn.

Closes GH-668.