Fail early when configuration file can't be found or is invalid.
It's confusing when you provide a configuration file using --config
option and it changes nothing. Usually this happens because JSHint
couldn't find your file or the file is not a valid JSON. This commit
makes JSHint to loudly complain about such cases.

Note that this affects only cases when user explicitly provides
a configuration file using --config.

Fixes GH-691.