Normalized stubbed .jshintrc paths for cli tests.
The paths that were being used to stub .jshintrc files did not match the
paths that would be searched for by the findConfig() function because
they are using hard coded forward slashes as path separators, but the
findConfig() function runs the supplied file path through path.resolve()
which returns normalized paths. This commit normalizes the stubbed
.jshintrc paths to match the functionality in findConfig() and support
windows systems.