Add a relaxing option to disable the unsafeChars check (W100)
This patch modifies the check for unsafe characters (unsafeChars) by
providing an option to relax it (option.unsafechars).

Several of the characters in the "unsafeChar" list are legitimate
Unicode characters (such as characters for the Sinahala language), and
can work properly if expected.

By default, this patch has no effect. It only causes a change in
behavior if the 'unsafechars' option is set to true.

References: GH-901