Fix Issue #1194
In Documentation/Configuration, first paragraph, is stated that:
"JSHint will look for this file in the current working directory and, if not
found, will move one level up the directory tree all the way up to the
filesystem root." (note: "this file" = ".jshintrc")
Ref: http://www.jshint.com/docs/config/

But this is not true if input is coming from stdin.
This patch address this unexpected behaviour by calling function findConfig
even when the input is coming from stdin.