Implemented 'safeasi' option
When using Automatic Semicolon Insertion there are
certain edge-cases that you need to keep in mind.
Namely new lines starting with either `(`, `[` or
`/` because those may mistakenly be interpreted as
continuations of the previous line.

If you enable the 'safeasi' option jshint will
watch out for all the potentially unsafe
edge-cases, so you don't have to.