Remove overzealous W092 (wrap regexp in parens after return stmt)
This warning flags any bare regexp after a return or yield statment as
ambiguous. (Example: `return /[0-9]/`.) The unnecessary fix is to wrap
the regexp in parens like so:

    return (/[0-9]/)

This is a useless warning. An unwrapped regexp following a return
statement is not ambiguous to humans or parsers. The division operator
(/) must follow an expression; a regexp literal must *not* follow an
expression. This warning flags regexp literals that appear immediately
after a return statement and thus never after an expression.

No reason to optionalize this warning, as it's not useful.

References:
  http://groups.yahoo.com/neo/groups/jslint_com/conversations/topics/345
  nzakas/eslint#324
  jshint/jshint#245
  jshint/jshint#1218