Update cli.js
Hello and thank you for JSHint,
At $work we have home directories on an automount path, and so for example with /home/users/slebedev/project_dir, if the user does not have a .jshintignore file in the cwd, directories above or home directory then /home/.jshintignore is stat'ed which for us triggers an automount action (ldap+NFS), while this should not cause any problems, it is something I noticed while investigating an automount issue.  I was just wondering if findFile in cli.js needs to search up beyond the home directory all the way to root?
Thanks.