Make 'forin' check recognize negative checks with continue
This patch modifies the 'forin' check so that it correctly recognizes
negative conditional checks using continue. For example:

for (var key in collection) {
    if (! collection.hasOwnProperty(key)) {
        continue;
    }
    doSomething();
}

The conditional itself is not analyzed beyond checking whether it
starts with a negation operator or not. Results of the if analysis
are stored in a stack that is evaluated after the for loop was parsed
completely. As a result, the improved check also works for nested loops.

References:
    Closes GH-229