[[FIX]] Remove dead code
The `_.has` utility method checks for the existence of the provided
properties as "own" values of the given object. The removed code
erroneously uses it to check attributes of an Array. The first branch is
(incorrectly) token when statements begin with identifiers/literals "0",
"1", or "length", and the nested branch depends on the block type which
is never a destructuring assignment in this case.

This code has been unreachable since its original introduction into the
codebase [1]. Removing it has no effect on the behavior of JSHint.

[1] commit fbf34a576f965aba3affda21e4e4507800162f57
    Author: Guyzmo <guyzmo+github@m0g.net>
    Date:   Wed Mar 6 15:12:15 2013 +0100

    fixed destructuring assign in global context. Updated tests. cf #883