[[FEAT]] Add enforcing option: `trailingcomma`
Missing trailing commas in array and object literals can result in
undetectable visual ambiguities--the same visual ambiguities that lead
to bugs when a programmer relies on automatic semicolon insertion.
Therefore, it might be a good practice to always use them, so this
option provides warnings in their absence.

For example, this code might have worked last Tuesday:

    [
        b + c
    ].forEach(print);

But if one adds an element to the array and forgets to compensate for
the missing comma, no syntax error is thrown, and a linter cannot
determine if this was a mistake or an intentional function invocation:

    [
        b + c
        (d + e)
    ].forEach(print);

If one always appends a list item with a comma, this ambiguity cannot
occur:

    [
        b + c,
    ].forEach(print);

    [
        b + c,
        (d + e),
    ].forEach(print);

If this option is `true`, a warning will be thrown if a "," does not
occur before the closing "]" or "}" in an array or object literal. If
option `es3` is also `true`, this option does nothing, because otherwise
the options would contradict each other, and the `es3` mandate is more
pressing.

There is no inverse of this option, that would warn if an ES5 programmer
omitted a trailing comma. Omission of a trailing comma in ES5 provides
no benefit, unlike option `trailingcomma`, which actually helps prevent
bugs. Such an inverse feature would be considered "stylistic;" do not
confuse `trailingcomma` as a "style" option, as it is no more
"stylistic" than the default detection of semicolons is.