[[FIX]] Correct behavior of `singleGroups`
Parsing logic can generally infer if the current token begins an
expression by referencing the `left` property of the previous token. The
`singleGroups` option relies on this behavior in order to recognize
situations where operator binding power makes the grouping operator
necessary.

Parsing of the "addition" operator should be made consistent with the
other infix operators: the token's `left` property should be set *prior
to* parsing the right-hand-side expression.