[[FEAT]] Parse and lint JSX
JSX is a syntactical addition spec'd here: http://facebook.github.io/jsx/. It's frequently used alongside the React view rendering library, but is also used in other projects.

To avoid needing to transform before linting, this adds the ability to parse and lint JSX laden code directly, provided that the `jsx` option is turned on.

Lint rules for undef follow the semantics of React if you ask it to by passing "react" to the "jsx" option, otherwise no assumptions are made about scoped variable usage.