[[FIX]] Do not crash on improper use of `delete`
In valid JavaScript code, when the `delete` token is parsed as a prefix,
it will be followed by an expression of some kind. Invalid code may
include any sequence of characters after the `delete` token, and JSHint
should be capable of gracefully warning about the syntax error.

Avoid crashing in cases where the `delete` appears as a "prefix" token
but is not followed by an expression.