[[FIX]] Remove tautological condition
The parsing logic for the `(` prefix token returns early if the token is
determined to mark the beginning of an arrow function parameter list.
Logic that follows can safely assume that the token is *not* part of an
arrow function parameter lists, and explicit conditions to verify this
are unnecessary.