[[FIX]] Refactor `doFunction`
The behavior of the `doFunction` routine is highly dynamic and dependent
on a number of optional arguments. None of these arguments is required,
and invocations of the function have become inconsistent in how these
are specified.

Re-factor `doFunction` to accept a single `options` argument. This
avoids the need to specify null values for irrelevant positional
arguments and makes the call site more legible to human readers.
Additionally, provide documentation for each of the options.