[[FIX]] Disambiguate argument
In order to operate for expressions and statements, the `doFunction`
routine requires only a boolean value describing whether the function
under construction is a statement or expression. The previous usage of a
positional argument named `statement` made this unclear, and calling
code differed in its usage.

Rename the option to `isStatement` to more clearly describe its expected
type and update calling code to consistently specify a boolean value.