[[FIX]] Permit "eval" as object key
Makes it possible to use "eval" as a key in an object. Previously this
would have emitted a "W061" warning. "eval" used as the key for a
presumed global ("window" in "browser:true" and "global" in
"node:true") will still emit the warning.

Fixes https://github.com/jshint/jshint/issues/738

Adds unit tests for permitting using "eval" as object key