[[FIX]] Don't treat referenced globals as locally defined
In this example, `foo` would be considered already defined when reaching
the `var` statement in the function:

    var foo;
    function x() {
      foo();
      var foo;
    }

This is because after parsing the first reference to foo it would appear
in the `funct` scope as "global". The `addlabel` function would then
treat it as already defined instead of simply referenced.

By handling global variables with the same branch as referenced
variables in `addlabel`, no warning for redefinition is generated, which
is the expected behavior.

Fixes #2138