[[FIX]] Prevent beginning array from being confused for JSON
src/jshint.js was reading an array at the beginning of a file as JSON instead
of a code block. The parser read the '(begin)' token and then looped 15 times
looking for an end to the block, didn't find it, and considered itself to be
parsing JSON. JSON must use *only* double quotes to be valid, so single quotes
caused warnings.