[[FIX]] Make let variables in the closure shadow predefs
There was a problem when parsing an identifier that would resolve to a
predefined read-only global through the function scope chain.

When checking the block scope for the same identifier only the innermost
block was considered, so while this worked:

    /* global foo: false */
    function x() {
      let foo;
      foo = {};
    }

... a W020 read only warning was generated when moving the assignment
into a subscope:

    /* global foo: false */
    function x() {
      let foo;
      {
        foo = {};
      }
    }

By checking the entire block scope chain for the identifier the
assignment generates no warning, as is expected.