Allow `this` in class expression methods
Commit d75ef694385ba44ab49b2d2b2cf1f4bed0949ca7 oversimplified the
information passed to the `doFunction` method.

Prior to that commit, `doFunction` was invoked inconsistently (with a
token and with an object literal). This was incorrectly simplified to a
boolean value, which caused the helper function `isMethod` to stop
working:

    function isMethod() {
      return funct["(statement)"] && funct["(statement)"].type === "class" ||
             funct["(context)"] && funct["(context)"]["(verb)"] === "class";
    }

Re-implement the option to be a token and specify consistently at each
invocation of `doFunction`.