[[Fix]] Only accept "symbol" as a type in ES6 envs
The ES6 specification extended the `typeof` operator to return the
string "symbol" when when applied to Symbol instances. In pre-ES6
environments, the operator will never return this value, so comparisons
that involve it cannot possibly evaluate to `true`.

Update the internal `isTypoTypeof` function to allow a different set of
values for different language versions.