[[FIX]] Relax restrictions on `singleGroups`
These operators are not specified as associative, so a grouping operator
may be necessary when they appear in a larger expression.

Update the binding power of these operators to accurately reflect the
way they are evaluated at runtime.

Remove invalid test cases that fail as a result of this correction.