[[FIX]] Incorrect 'Unclosed string' when the closing quote is the first character after a newline
Previously, the end-of-line loop was immediately followed by the non-quote character code. This
meant that the first character after an EOL was always treated as a non-quote character.

Control flow now returns to the quote character check in the while loop condition immediately
after each EOL or character has been handled, so that this check also runs for the first character
on a line.

Closes #1532. (This commit includes the test case provided in #1532 by DelvarWorld.)
Closes #1319.