[[FIX]] Prevent incorrect warnings for relations
Denote relation operator tokens as binary operators by attaching a
reference to the left-hand expression *prior* to parsing the right-hand
expression.

This allows the logic for the `singleGroups` option to correctly infer
that a grouping operator does not begin an expression when it follows a
relation operator. This in turn ensures that the `singleGroups` option
does not trigger warnings when the grouping operator is necessary and
involves a relation operator.