[[FIX]] Distinguish between directive and mode
All code within a class body is interpreted as "strict mode" code. This
can be simulated by explicitly setting the internal mechanism for the
"use strict" directive, but that approach is misleading. In such cases,
the code is in strict mode, but the "use strict" directive is *not* set.

Define a general-purpose `isStrict` method that determines the language
mode according to these two distinct conditions.

This approach enables a more expressive and accurate means for the
internals to control parsing behavior. In addition, it may be re-used
when JSHint is extended to lint module bodies as strict mode code.