[[FIX]] Move helper methods to `state` object
Declare these helper functions as part of the definition of the `state`
object. This makes these methods easier to discover. There is no need to
re-instantiate the methods when the state object is re-setand, so this
approach is slightly more efficient.

Remove unnecessary parameters and replace necessary boolean parameter
with an options object.