[[FEAT]] Implement `module` option
As per the ECMAScript 6 specification, all module code is always strict
code. This option allows users to signal when the input code is intended
to be interpreted as a JavaScript module. This enables JSHint's
strict-mode-specific functionality without requiring a "use strict"
directive appear in the source (as this is redundant in compliant
runtimes).

Introduce a new error to restrict this option to code written as
ECMAScript 6 or later.