[[Fix]] Update Lodash
When the 3.6.0 release of Lodash is included in a script using
Browserify, the resulting file cannot be executed in a browser web
worker context. This is how the jshint.com homepage includes JSHint,
meaning that the website cannot be updated to demonstrate the latest
release.

The 3.7.0 release of Lodash addresses this problem [1].

Update to Lodash version 3.7.0 so the next release can be published on
the JSHint homepage.

[1]
https://github.com/lodash/lodash/commit/e28e04a99030ef968a2ab1f9b3d9ed7ea01b7797