[[FIX]] Relax singleGroups restrictions: IIFEs
Although not technically necessary, the grouping operator is commonly
used to "wrap" function expressions in cases where it is the left-hand
side of some operator (typically as the target of invocation). This
pattern affords readers a hint that the function object in question is
the target of some further operation--likely many lines after the
initial `function` keyword.

Update the behavior of the `singleGroups` option to allow for such
usages since they serve a beneficial purpose beyond modifying expression
execution order.