[[FIX]] Accept new RegExp flag introduced by ES6
ECMAScript 6 introduces the "u" flag for regular expressions. Extend the
lexer to accept this when linting ES6 code.

Existing code attempts to compile regular expression literals as a
mechanism to ensure validity. Update this code to ignore regular
expression flags in order to ensure consistent behavior when JSHint is
executed in ES5 and ES6 environments.