[[FEAT]] Support .jshintrc "extends" with no specific path.
Extending the abilities of the .jshintrc option "extends".
The new behaviour resembles the "root" option of .editorconfig files.

When the boolean value true is used:
JSHint will walk up the directory tree and look for a .jshintrc file.

If no .jshintrc file is found (root directory reached):
Nothing will happen, and the original .jshintrc file will be used as is.

If a .jshintrc file is found:
JSHint will behave as if the file's path was explicitly paseed as the "extends" option.
This includes recursive behavior for this feature.